Basic implementation

+++ CompleteGPSUtility/Database/Models/LocationModel.cs
@@ -4,7 +4,7 @@
     {
         public decimal a { get; set; }                  //Latatitude decimal
         public decimal o { get; set; }                  //Longitude decimal
-        public decimal? l { get; set; }                 //Altitude in meters
+        public short? l { get; set; }                 //Altitude in meters
         public int y { get; set; }                      //Y2K = Seconds from 2000-01-01 in UTC
     }
 }
